// HRHelper-frontend/src/types/dashboardDTO.ts
export interface HoursSummary {
    regular: number;
    overtimeDay: number;
    overtimeNight: number;
    overtimeHoliday: number;
    leaveVacation: number;
    leaveUnpaid: number;
    leaveCircumstance: number;
    leavePregnant: number;
    leaveParental: number;
    leaveTraining: number;
    leaveHigherPower: number;
    leaveJobSearch: number;
    leaveBlood: number;
    leaveCarer: number;
    sickLeave: number;
  }
  